move DCOPY_statistics to be internal to mfu_flist_copy
